refactor: ♻️ Refector token

+++ src/merchants/merchants.controller.ts
@@ -6,7 +6,10 @@ import {
   Param,
   Patch,
   Post,
+  Request,
+  UseGuards,
 } from '@nestjs/common';
+import { AuthGuard } from '@nestjs/passport';
 import {
   ApiBadRequestResponse,
   ApiBearerAuth,
@@ -26,6 +29,7 @@ import { MerchantsService } from './merchants.service';
 @ApiBearerAuth()
 @ApiTags('Merchants')
 @Controller('merchants')
+@UseGuards(AuthGuard())
 export class MerchantsController {
   constructor(private readonly merchantsService: MerchantsService) {}
 
@@ -38,8 +42,8 @@ export class MerchantsController {
   @ApiForbiddenResponse({ description: 'Forbidden.' })
   @ApiBody({ type: Merchant })
   @ApiOperation({ summary: 'Cria uma novo merchant' })
-  create(@Body() createMerchantDto: CreateMerchantDto) {
-    return this.merchantsService.create(createMerchantDto);
+  create(@Request() req, @Body() createMerchantDto: CreateMerchantDto) {
+    return this.merchantsService.create(createMerchantDto, req.user.id);
   }
 
   @Get()
@@ -50,20 +54,20 @@ export class MerchantsController {
   })
   @ApiForbiddenResponse({ description: 'Forbidden.' })
   @ApiOperation({ summary: 'Retorna o merchant do usuário' })
-  findAll() {
-    return this.merchantsService.findAll();
+  findAll(@Request() req) {
+    return this.merchantsService.findAll({ userId: req.user.id });
   }
 
-  @Get(':id')
-  @ApiOperation({ summary: 'Retorna o merchant do usuário pelo ID' })
-  @ApiOkResponse({
-    description: 'Retorna o merchant do usuário',
-    type: Merchant,
-  })
-  @ApiForbiddenResponse({ description: 'Forbidden.' })
-  findOne(@Param('id') id: string) {
-    return this.merchantsService.findOne(+id);
-  }
+  // @Get(':id')
+  // @ApiOperation({ summary: 'Retorna o merchant do usuário pelo ID' })
+  // @ApiOkResponse({
+  //   description: 'Retorna o merchant do usuário',
+  //   type: Merchant,
+  // })
+  // @ApiForbiddenResponse({ description: 'Forbidden.' })
+  // findOne(@Param('id') id: string) {
+  //   return this.merchantsService.findOne(+id);
+  // }
 
   @Patch(':id')
   @ApiOperation({ summary: 'Atualiza merchant do usuário pelo ID' })
@@ -75,10 +79,14 @@ export class MerchantsController {
   @ApiBadRequestResponse({ description: 'Bad Request' })
   @ApiForbiddenResponse({ description: 'Forbidden.' })
   update(
+    @Request() req,
     @Param('id') id: string,
     @Body() updateMerchantDto: UpdateMerchantDto,
   ) {
-    return this.merchantsService.update(+id, updateMerchantDto);
+    return this.merchantsService.update(
+      { id, userId: req.user.id },
+      updateMerchantDto,
+    );
   }
 
   @Delete(':id')
@@ -86,7 +94,7 @@ export class MerchantsController {
   @ApiOkResponse({ description: 'O merchant foi removido com sucesso' })
   @ApiForbiddenResponse({ description: 'Forbidden.' })
   @ApiNotFoundResponse({ description: 'Profile Not found' })
-  remove(@Param('id') id: string) {
-    return this.merchantsService.remove(+id);
+  remove(@Request() req, @Param('id') id: string) {
+    return this.merchantsService.remove({ id, userId: req.user.id });
   }
 }
